fft,ifft빼고 모두 int형으로 고친것

+++ 졸작/decodertoc/AwgnGen.h
@@ -6,7 +6,7 @@
 
 float Normal(float x, float sd);
 // AWGN 
-int AwgnGen(float *input_buffer, float *output_buffer
+int AwgnGen(int *input_buffer, int *output_buffer
 			, int buffer_length, float EbNo)
 {
 	int i;
